Adding NaturalField facility, Sunflower plant, and IFlower interface.…

+++ src/Interfaces/ISeedProducing.cs
@@ -2,6 +2,6 @@ namespace Trestlebridge.Interfaces
 {
     public interface ISeedProducing
     {
-        double Harvest ();
+        int Harvest ();
     }
 }
